refactored. All Banners fixed. One API divided into two separated
