--- src/components/Main/HotPricesList/index.ts
@@ -1,1 +1,0 @@
-export * from './HotPricesList';
