--- src/components/Main/NewModelsList/index.ts
@@ -1,1 +1,0 @@
-export * from './NewModelList';
