+++ src/components/ProductDetails/RecommendedList/index.tsx
@@ -1,0 +1,1 @@
+export * from './RecommendedList';
