+++ src/pages/ProductDetailsPage/ProductDetailsPage.tsx
@@ -1,38 +1,30 @@
-import { useEffect, useMemo, useState } from 'react';
+import { useEffect, useState } from 'react';
 import { useLocation, useParams } from 'react-router-dom';
 import styles from './ProductDetailsPage.module.scss';
 import { Loading } from '../../components/Loading';
 import { Breadcrumbs } from '../../components/Breadcrumbs';
 import { About } from '../../components/ProductDetails/About';
-import { GoodsSlider } from '../../components/Main/GoodsSlider';
 import { Gallery } from '../../components/ProductDetails/Gallery';
 import { TechSpecs } from '../../components/ProductDetails/TechSpecs';
 import { VariantsActions } from '../../components/ProductDetails/VariantsActions';
 import { UnionProduct } from '../../utils/types/UnionProduct';
 import { ProductCategory } from '../../utils/types/ProductCategory';
 import { BackButton } from '../../components/Buttons/BackButton';
-import { chooseProductAPI } from '../../utils/helpers/chooseProductAPI';
+import { getProduct } from '../../api/api';
+import { RecommendedList } from '../../components/ProductDetails/RecommendedList';
 
 export const ProductDetailsPage = () => {
   const [product, setProduct] = useState<UnionProduct | null>(null);
-  const [currentProductList, setCurrentProductList] = useState<UnionProduct[]>(
-    [],
-  );
   const [isLoading, setIsLoading] = useState(true);
   const [error, setError] = useState('');
   const { pathname } = useLocation();
   const { productId } = useParams() as { productId: string };
   const hasBackButton = window.history.length > 1;
 
-  const getProduct = useMemo(() => {
-    return chooseProductAPI(pathname);
-  }, [pathname]);
-
   useEffect(() => {
-    getProduct()
-      .then(products => {
-        const data = products.find(product => product.id === productId);
-        setCurrentProductList(products);
+    getProduct(pathname as string, productId as string)
+      .then(product => {
+        const data = product;
 
         if (data) {
           document.title = data.name;
@@ -48,7 +40,7 @@ export const ProductDetailsPage = () => {
       .finally(() => {
         setIsLoading(false);
       });
-  }, [setIsLoading, setProduct, productId]);
+  }, [setIsLoading, setProduct, productId, pathname]);
 
   if (isLoading) {
     return (
@@ -119,10 +111,7 @@ export const ProductDetailsPage = () => {
         </div>
       </div>
 
-      <GoodsSlider
-        sliderTitle="You may also like"
-        productsList={currentProductList}
-      />
+      <RecommendedList />
     </div>
   );
 };
